Reduce to 8 bytes per tile for GPU
x, y, z, tex coord id each 2 bytes

+++ include/core/data_type.h
@@ -34,7 +34,7 @@ namespace jactorio
     using SpriteSetT           = uint16_t;
     using SpriteFrameT         = uint16_t;
     using SpriteTrimT          = uint16_t;
-    using SpriteTexCoordIndexT = uint32_t;
+    using SpriteTexCoordIndexT = uint16_t;
 
     using ResourceEntityResourceCount = uint32_t;
